requirements.txt
streamlit
pyyaml
pandas
plotly
scikit-learn
openai
nltk
python-dotenv
